Make highlights utility reusable
