+++ client/my-sites/stats/wordads/highlights-section.jsx
@@ -1,95 +1,15 @@
 import { Popover } from '@automattic/components';
-import { Icon, info, payment, receipt, tip } from '@wordpress/icons';
-import { numberFormat, translate } from 'i18n-calypso';
+import { Icon, info } from '@wordpress/icons';
+import { translate } from 'i18n-calypso';
 import { useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { getWordAdsEarnings } from 'calypso/state/wordads/earnings/selectors';
 import HighlightCardSimple from './highlight-card-simple';
+import { getEarningsSummaries, getPayoutNotices, highlightIconById } from './utils';
 
 // TODO: HighlightCard does not accept string values.
 // Should refactor to accept strings and move the business logic into the callers.
 // Then refactor this Comp to use HighlightCard again.
-
-function getAmountAsFormattedString( amount ) {
-	if ( amount === 0 ) {
-		// Per design spec we don't want "$0.00" as a result.
-		// https://github.com/Automattic/wp-calypso/issues/72045
-		// We'll return "$0" in this scenario.
-		return '$0';
-	}
-	// Takes a Number, formats it to 2 decimal places, and prepends a "$".
-	// Amounts are in USD with localized formatting.
-	return '$' + numberFormat( amount, 2 );
-}
-
-function getHighlights( earnings ) {
-	const total = earnings?.total_earnings ? Number( earnings.total_earnings ) : 0;
-	const owed = earnings?.total_amount_owed ? Number( earnings.total_amount_owed ) : 0;
-	const paid = total - owed;
-
-	const highlights = [
-		{
-			heading: translate( 'Earnings', { comment: 'Total WordAds earnings to date' } ),
-			icon: <Icon icon={ payment } />,
-			amount: total,
-		},
-		{
-			heading: translate( 'Paid', {
-				comment: 'Total WordAds earnings that have been paid out',
-			} ),
-			icon: <Icon icon={ receipt } />,
-			amount: paid,
-		},
-		{
-			heading: translate( 'Outstanding amount', {
-				comment: 'Total WordAds earnings currently unpaid',
-			} ),
-			icon: <Icon icon={ tip } />,
-			amount: owed,
-		},
-	];
-	// Transform/index the data for use with React.
-	return highlights.map( ( highlight, i ) => ( {
-		id: i,
-		formattedAmount: getAmountAsFormattedString( highlight.amount ),
-		...highlight,
-	} ) );
-}
-
-function payoutNotices( earnings ) {
-	const amountOwed = earnings?.total_amount_owed || 0;
-	const amountOwedFormatted = getAmountAsFormattedString( amountOwed );
-	const notice = {
-		id: 'notice',
-		value: translate(
-			'Outstanding amount of %(amountOwed)s does not exceed the minimum $100 needed to make the payment.',
-			{
-				comment: 'WordAds: Insufficient balance for payout.',
-				args: { amountOwed: amountOwedFormatted },
-			}
-		),
-	};
-	const limit = {
-		id: 'limit',
-		value: translate(
-			'Payment will be made as soon as the total outstanding amount has reached $100.',
-			{
-				comment: 'WordAds: Payout limit description.',
-			}
-		),
-	};
-	const payout = {
-		id: 'payout',
-		value: translate(
-			'Outstanding amount of %(amountOwed)s will be paid approximately 45 days following the end of the month in which it was earned.',
-			{
-				comment: 'WordAds: Payout will proceed.',
-				args: { amountOwed: amountOwedFormatted },
-			}
-		),
-	};
-	return amountOwed < 100 ? [ notice, limit ] : [ payout ];
-}
 
 function HighlightsSectionHeader( props ) {
 	const [ isTooltipVisible, setTooltipVisible ] = useState( false );
@@ -136,7 +56,7 @@ function HighlightsListing( props ) {
 				<HighlightCardSimple
 					key={ highlight.id }
 					heading={ highlight.heading }
-					icon={ highlight.icon }
+					icon={ <Icon icon={ highlightIconById[ highlight.id ] } /> }
 					value={ highlight.formattedAmount }
 				/>
 			) ) }
@@ -146,8 +66,8 @@ function HighlightsListing( props ) {
 
 export default function HighlightsSection( props ) {
 	const earningsData = useSelector( ( state ) => getWordAdsEarnings( state, props.siteId ) );
-	const highlights = getHighlights( earningsData );
-	const notices = payoutNotices( earningsData );
+	const highlights = getEarningsSummaries( earningsData );
+	const notices = getPayoutNotices( earningsData );
 	return (
 		<div className="highlight-cards wordads has-background-color">
 			<HighlightsSectionHeader notices={ notices } />
